[FIX MASTER] fix runtime -> now fix line of wall

+++ sources/graphical/sprite_draw.c
@@ -6,7 +6,7 @@
 /*   By: lfallet <lfallet@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/06/10 12:23:14 by lfallet           #+#    #+#             */
-/*   Updated: 2020/06/23 22:26:30 by lfallet          ###   ########.fr       */
+/*   Updated: 2020/06/24 12:09:37 by lfallet          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -58,11 +58,11 @@ void	calcdraw_sprite(t_graph *gr, t_map *map)
 	gr->sp.endy = gr->sp.height / 2 + map->recup.resolution[AXE_Y] / 2;
 	if (gr->sp.endy >= map->recup.resolution[AXE_Y])
 		gr->sp.endy	= map->recup.resolution[AXE_Y] - 1;
-	gr->sp.width = abs((int)(map->recup.resolution[AXE_Y] / gr->sp.playposy));
-	gr->sp.startx = -gr->sp.width / 2 + gr->sp.screen_where;
+	gr->sp.width = abs((unsigned int)(map->recup.resolution[AXE_Y] / gr->sp.playposy));
+	gr->sp.startx = (unsigned int)(-gr->sp.width / 2 + gr->sp.screen_where);
 	if (gr->sp.startx < 0)
 		gr->sp.startx = 0;
 	gr->sp.endx = gr->sp.width / 2 + gr->sp.screen_where;
 	if (gr->sp.endx >= map->recup.resolution[AXE_X])
-			gr->sp.endx = map->recup.resolution[AXE_X];	
+			gr->sp.endx = map->recup.resolution[AXE_X];
 }
